Move components to src/components/auth, add index.ts exports
