+++ packages/react/spec/components/auth/SignedOut.spec.tsx
@@ -1,10 +1,10 @@
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
 import React from "react";
-import { superAuthApi } from "../../spec/apis";
-import { TestWrapper } from "../../spec/testWrapper";
-import { expectMockSignedInUser, expectMockSignedOutUser } from "../../spec/utils";
-import { SignedOut } from "../../src/components/SignedOut";
+import { superAuthApi } from "../../../spec/apis";
+import { TestWrapper } from "../../../spec/testWrapper";
+import { expectMockSignedInUser, expectMockSignedOutUser } from "../../../spec/utils";
+import { SignedOut } from "../../../src/components/auth/SignedOut";
 
 describe("SignedOut", () => {
   test("renders children when signed out", () => {
