+++ packages/react/src/auth/utils.ts
@@ -1,7 +1,7 @@
 import type { GadgetSession } from "./useSession";
 
 export const isSessionSignedOut = (session: GadgetSession) => {
-  return session?.userId == undefined;
+  return !session?.userId;
 };
 
 export const isSessionSignedIn = (session: GadgetSession) => {
