+++ packages/react/src/components/auth/SignedIn.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import React from "react";
-import { useSession } from "../../src/auth/useSession";
-import { isSessionSignedIn } from "../../src/auth/utils";
+import { useSession } from "../../auth/useSession";
+import { isSessionSignedIn } from "../../auth/utils";
 
 /**
  * Renders its `children` if the current `Session` is signed in (has an associated `User`), otherwise renders nothing.
