+++ packages/react/src/components/auth/SignedInOrRedirect.tsx
@@ -1,8 +1,8 @@
 import type { ReactNode } from "react";
 import React, { useContext, useEffect, useState } from "react";
-import { GadgetClientContext } from "../../src/GadgetProvider";
-import { useSession } from "../../src/auth/useSession";
-import { isSessionSignedIn } from "../../src/auth/utils";
+import { GadgetClientContext } from "../../../src/GadgetProvider";
+import { useSession } from "../../../src/auth/useSession";
+import { isSessionSignedIn } from "../../../src/auth/utils";
 
 /**
 /**
